AAH_ASynchronousCommunicator is in 1 .h/.cc file
