+++ framework/mesh/SweepUtilities/Communicators/AAH_AsynComm.h
@@ -17,9 +17,9 @@ namespace chi_mesh::sweep_management
 
 class FLUDS;
 
-// ###################################################################
-/**Handles the swift communication of interprocess communication
- * related to sweeping.*/
+/**
+ * Handles the swift communication of interprocess communication related to sweeping.
+ */
 class AAH_ASynchronousCommunicator : public AsynchronousCommunicator
 {
 private:
@@ -57,17 +57,71 @@ public:
                                size_t num_angles,
                                int sweep_eager_limit,
                                const chi::ChiMPICommunicatorSet& in_comm_set);
+  /**
+   * Returns the private flag done_sending.
+   */
   bool DoneSending() const;
+
+  /**
+   * Initializes delayed upstream data. This method gets called
+   * when a sweep scheduler is constructed.
+   */
   void InitializeDelayedUpstreamData();
+
+  /**
+   * This is the final level of initialization before a sweep-chunk executes.
+   * Once all upstream dependencies are met and if the sweep scheduler places
+   * this angleset as "ready-to-execute", then the angle-set will call this
+   * method. It is also fairly important in terms of memory to only allocate
+   * these chunks of memory when actually ready to use them since they form the
+   * majority of memory usage.
+   */
   void InitializeLocalAndDownstreamBuffers();
+
+  /**
+   * Sends downstream psi. This method gets called after a sweep chunk has executed
+   */
   void SendDownstreamPsi(int angle_set_num);
+
+  /**
+   * Receives delayed data from successor locations.
+   */
   bool ReceiveDelayedData(int angle_set_num);
+
+  /**
+   * Sends downstream psi.
+   */
   void ClearDownstreamBuffers();
+
+  /**
+   * Check if all upstream dependencies have been met and receives
+   * it as it becomes available.
+   */
   AngleSetStatus ReceiveUpstreamPsi(int angle_set_num);
+
+  /**
+   * Receive all upstream Psi. This method is called from within
+   * an advancement of an angleset, right after execution.
+   */
   void ClearLocalAndReceiveBuffers();
+
+  /**
+   * Clear flags in preperation for another sweep.
+   */
   void Reset();
 
 protected:
+  /**
+   * Builds message structure.
+   *
+   * Outgoing and incoming data needs to be sub-divided into messages
+   * each of which is smaller than the MPI eager-limit. There are
+   * three parts to this: predecessors, delayed-predecessors and successors.
+   *
+   * This method gets called by an angleset that subscribes to this
+   * sweepbuffer.
+   */
   void BuildMessageStructure();
 };
+
 } // namespace chi_mesh::sweep_management
